Add projection to watch and map max once/sec, varsToWatch to model
SysModModel: move varsToWatch here (from E131)
AppModLeds: 
- remove E131 dependency, 
- add projection
- add doMap and lastMappingMillis
- call ledFixProjectAndMap in loop (max 1/sec)
