+++ src/User/UserModE131.h
@@ -13,19 +13,9 @@
 
 #include <vector>
 
-#define maxChannels 10
-
-struct VarToWatch {
-  const char * id = nullptr;
-  uint16_t max = -1;
-  uint8_t savedValue = -1;
-};
-
 class UserModE131:public Module {
 
 public:
-
-  VarToWatch varsToWatch[maxChannels]; //up to 513
 
   UserModE131() :Module("e131/sACN support") {
     print->print("%s %s\n", __PRETTY_FUNCTION__, name);
@@ -69,7 +59,7 @@ public:
         e131.pull(&packet);     // Pull packet from ring buffer
 
         for (int i=0; i < maxChannels; i++) {
-          if (packet.property_values[i] != varsToWatch[i].savedValue) {
+          if (packet.property_values[i] != mdl->varsToWatch[i].savedValue) {
 
             print->print("Universe %u / %u Channels | Packet#: %u / Errors: %u / CH%d: %u -> %u",
                     htons(packet.universe),                 // The Universe for this packet
@@ -77,25 +67,20 @@ public:
                     e131.stats.num_packets,                 // Packet counter
                     e131.stats.packet_errors,               // Packet error counter
                     i,
-                    varsToWatch[i].savedValue,
+                    mdl->varsToWatch[i].savedValue,
                     packet.property_values[i]);             // Dimmer data for Channel i
 
-            varsToWatch[i].savedValue = packet.property_values[i];
+            mdl->varsToWatch[i].savedValue = packet.property_values[i];
 
-            if (varsToWatch[i].id != nullptr) {
-              print->print(" var: %s\n", varsToWatch[i].id);
-              mdl->setValueI(varsToWatch[i].id, varsToWatch[i].savedValue%varsToWatch[i].max); // TODO: ugly to have magic string 
+            if (mdl->varsToWatch[i].id != nullptr) {
+              print->print(" var: %s\n", mdl->varsToWatch[i].id);
+              mdl->setValueI(mdl->varsToWatch[i].id, mdl->varsToWatch[i].savedValue%mdl->varsToWatch[i].max); // TODO: ugly to have magic string 
             }
             else
               print->print("\n");
           }
         }
     }
-  }
-
-  void addWatch(uint8_t channel, const char * id, uint16_t max) {
-    varsToWatch[channel].id = id;
-    varsToWatch[channel].max = max;
   }
 
   private:
